+++ NEWS.md
@@ -1,15 +1,17 @@
 # poldis 0.1.2
 
-2024-08-08
+2024-09-03
 
 ## Package
 
 - Updated documentation for `get_urgency()` to include more details about urgency scores
+- Updated internal data and added a data raw folder to increase transparency
 
 ## Functions
 
 - Updated `get_urgency()` and `gather_topics()` to open codebooks for urgency or topic when no arguments are declared
 - Updated urgency scores in `get_urgency()` according to survey results. 
+- Added `sim_urgency()` function to simulate urgent priorities
 
 # poldis 0.1.1